message_edit_history: Move click handlers to message_edit_history.js.
